Fixed an issue that made the program think a background opertaion was still running if it crashed

+++ MusicPlayer/MusicPlayer/UpdateExports.cs
@@ -44,7 +44,8 @@ namespace MusicPlayer
                 string name = Path.GetFileNameWithoutExtension(SongPaths[i]);
                 try
                 {
-                    File.Copy(SongPaths[i], Target + "\\" + Path.GetFileName(SongPaths[i]));
+                    if (!File.Exists(Target + "\\" + Path.GetFileName(SongPaths[i])))
+                        File.Copy(SongPaths[i], Target + "\\" + Path.GetFileName(SongPaths[i]));
                 }
                 catch (Exception ex)
                 {
